# packages/api-contracts/openapi.yaml
...
openapi: 3.0.0
paths:
...
  /api/v1/device-tokens:
    post:
      description: Requires authentication. Idempotent per token.
      operationId: DeviceTokenController_register_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  minLength: 10
                platform:
                  type: string
                  enum:
                    - android
                    - ios
                    - web
                  nullable: true
              required:
                - token
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/OkResultDto'
                required:
                  - success
                  - data
        '400':
          description: VALIDATION_ERROR — the request failed schema validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '401':
          description: AUTHENTICATION_ERROR — missing, invalid or expired bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      security:
        - bearer: []
      summary: Register an FCM device token for push
      tags:
        - device-tokens
    delete:
      description: Requires authentication. The token is sent in the request BODY, not the path.
      operationId: DeviceTokenController_unregister_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  minLength: 10
                platform:
                  type: string
                  enum:
                    - android
                    - ios
                    - web
                  nullable: true
              required:
                - token
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/OkResultDto'
                required:
                  - success
                  - data
        '400':
          description: VALIDATION_ERROR — the request failed schema validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '401':
          description: AUTHENTICATION_ERROR — missing, invalid or expired bearer token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      security:
        - bearer: []
      summary: Unregister a device token
      tags:
        - device-tokens
...
  /api/v1/discovery/trending:
    get:
      description: Public (guest). Derived from real logged activity; returns an empty list rather than padding when there is not enough data.
      operationId: DiscoveryController_trending_v1
      parameters:
        - name: limit
          required: false
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
          description: Maximum rows to return.
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TrendingResponseDto'
                  meta:
                    $ref: '#/components/schemas/CountMetaDto'
                required:
                  - success
                  - data
        '400':
          description: VALIDATION_ERROR — the request failed schema validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      summary: Trending products
      tags:
        - discovery
  /api/v1/discovery/most-searched:
    get:
      description: Public (guest). searchCount comes from logged search analytics over the requested window — never fabricated.
      operationId: DiscoveryController_mostSearched_v1
      parameters:
        - name: hours
          required: false
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 168
            default: 24
          description: Look-back window in hours.
        - name: limit
          required: false
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 12
          description: Maximum rows to return.
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MostSearchedProductDto'
                  meta:
                    $ref: '#/components/schemas/CountMetaDto'
                required:
                  - success
                  - data
        '400':
          description: VALIDATION_ERROR — the request failed schema validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      summary: Most-searched products in a recent window
      tags:
        - discovery
  /api/v1/discovery/most-searched/{category}:
    get:
      description: Public (guest). Same signal as /most-searched, scoped to a category slug.
      operationId: DiscoveryController_mostSearchedInCategory_v1
      parameters:
        - name: category
          required: true
          in: path
          description: Category slug.
          schema:
            type: string
        - name: hours
          required: false
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 168
            default: 24
          description: Look-back window in hours.
        - name: limit
          required: false
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 12
          description: Maximum rows to return.
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MostSearchedProductDto'
                  meta:
                    $ref: '#/components/schemas/CountMetaDto'
                required:
                  - success
                  - data
        '400':
          description: VALIDATION_ERROR — the request failed schema validation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '404':
          description: NOT_FOUND — the requested resource does not exist.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '429':
          description: Rate limit exceeded (global ThrottlerGuard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
        '500':
          description: INTERNAL_ERROR — unexpected failure; details are logged, never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponseDto'
      summary: Most-searched products within one category
      tags:
        - discovery
info:
  title: Omnisift API
  description: Product Intelligence Platform — REST API
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - AUTHENTICATION_ERROR
        - AUTHORIZATION_ERROR
        - NOT_FOUND
        - PROVIDER_ERROR
        - DATABASE_ERROR
        - BUSINESS_RULE_ERROR
        - INTERNAL_ERROR
      description: Stable machine-readable code. Safe to branch on.
    ApiErrorDetailDto:
      type: object
      properties:
        code:
          description: Stable machine-readable code. Safe to branch on.
          allOf:
            - $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Safe, user-facing message. Never contains a stack trace or internal detail.
      required:
        - code
        - message
    ApiErrorResponseDto:
      type: object
      properties:
        success:
          type: boolean
          example: false
          description: Always false on an error response.
        error:
          $ref: '#/components/schemas/ApiErrorDetailDto'
        requestId:
          type: string
          nullable: true
          description: Correlation id echoed from the x-request-id request header, or null.
      required:
        - success
        - error
        - requestId
...
    OkResultDto:
      type: object
      properties:
        ok:
          type: boolean
          example: true
      required:
        - ok
...
    CountMetaDto:
      type: object
      properties:
        count:
          type: number
          description: Number of items in `data`.
      required:
        - count
...
    TrendingResponseDto:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title:
          type: string
        imageUrl:
          type: string
          nullable: true
        lowestPrice:
          type: number
          nullable: true
        discountPct:
          type: number
          nullable: true
        sellerCount:
          type: number
      required:
        - id
        - slug
        - title
        - imageUrl
        - lowestPrice
        - discountPct
        - sellerCount
    MostSearchedProductDto:
      type: object
      properties:
        token:
          type: string
        productId:
          type: string
        title:
          type: string
        platform:
          type: string
        price:
          type: number
        oldPrice:
          type: number
          nullable: true
        currency:
          type: string
        imageUrl:
          type: string
        productUrl:
          type: string
        rating:
          type: number
          nullable: true
        reviewCount:
          type: number
          nullable: true
        delivery:
          type: string
        brand:
          type: string
        category:
          type: string
        searchCount:
          type: number
          description: Times this product was returned for a search in the requested `hours` window.
      required:
        - token
        - productId
        - title
        - platform
        - price
        - oldPrice
        - currency
        - imageUrl
        - productUrl
        - rating
        - reviewCount
        - delivery
        - brand
        - category
        - searchCount